Fix link synteny view
