+++ plugins/linear-comparative-view/src/LGVSyntenyDisplay/components/LaunchSyntenyViewDialog.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react'
 
 import { Dialog } from '@jbrowse/core/ui'
-import { getSession } from '@jbrowse/core/util'
 import {
   Button,
   Checkbox,
@@ -14,7 +13,8 @@ import { makeStyles } from 'tss-react/mui'
 
 import { navToSynteny } from './util'
 
-import type { Feature } from '@jbrowse/core/util'
+import type { AbstractSessionModel, Feature } from '@jbrowse/core/util'
+import type { LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
 
 const useStyles = makeStyles()({
   padding: {
@@ -24,23 +24,41 @@ const useStyles = makeStyles()({
 })
 
 export default function LaunchSyntenyViewDialog({
-  model,
+  session,
+  view,
   feature,
   trackId,
   handleClose,
 }: {
-  model: unknown
+  session: AbstractSessionModel
+  view?: LinearGenomeViewModel
   feature: Feature
   trackId: string
   handleClose: () => void
 }) {
   const { classes } = useStyles()
   const inverted = feature.get('strand') === -1
+  const hasCIGAR = !!feature.get('CIGAR')
   const [horizontallyFlip, setHorizontallyFlip] = useState(inverted)
   const [windowSize, setWindowSize] = useState('1000')
+  const [useRegionOfInterest, setUseRegionOfInterest] = useState(true)
   return (
     <Dialog open title="Launch synteny view" onClose={handleClose}>
       <DialogContent>
+        {view && hasCIGAR ? (
+          <FormControlLabel
+            className={classes.padding}
+            control={
+              <Checkbox
+                checked={useRegionOfInterest}
+                onChange={event => {
+                  setUseRegionOfInterest(event.target.checked)
+                }}
+              />
+            }
+            label="Use CIGAR string to navigate the current visible to the target"
+          />
+        ) : null}
         {inverted ? (
           <FormControlLabel
             className={classes.padding}
@@ -77,11 +95,14 @@ export default function LaunchSyntenyViewDialog({
                   windowSize: +windowSize,
                   horizontallyFlip,
                   trackId,
-                  model,
+                  session,
+                  region: useRegionOfInterest
+                    ? view?.dynamicBlocks.contentBlocks[0]
+                    : undefined,
                 })
               } catch (e) {
                 console.error(e)
-                getSession(model).notifyError(`${e}`, e)
+                session.notifyError(`${e}`, e)
               }
             })()
             handleClose()
