+++ plugins/linear-comparative-view/src/LGVSyntenyDisplay/components/util.ts
@@ -1,10 +1,7 @@
-import { getSession } from '@jbrowse/core/util'
 import { MismatchParser } from '@jbrowse/plugin-alignments'
 
 import type { LinearSyntenyViewModel } from '../../LinearSyntenyView/model'
-import type { Feature } from '@jbrowse/core/util'
-import type { LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
-import type { IAnyStateTreeNode } from 'mobx-state-tree'
+import type { AbstractSessionModel, Feature, Region } from '@jbrowse/core/util'
 
 type LSV = LinearSyntenyViewModel
 
@@ -36,23 +33,27 @@ function findPosInCigar(cigar: string[], startX: number) {
   return [featX, mateX] as const
 }
 
+export interface SimpleRegion {
+  refName: string
+  start: number
+  end: number
+}
+
 export async function navToSynteny({
   feature,
   windowSize: ws,
-  model,
+  session,
   trackId,
-  view,
+  region,
   horizontallyFlip,
 }: {
   windowSize: number
   trackId: string
   horizontallyFlip: boolean
   feature: Feature
-  view?: LinearGenomeViewModel
-  model: IAnyStateTreeNode
+  session: AbstractSessionModel
+  region?: SimpleRegion
 }) {
-  const session = getSession(model)
-  const reg = view?.dynamicBlocks.contentBlocks[0]
   const cigar = feature.get('CIGAR')
   const strand = feature.get('strand')
 
@@ -71,9 +72,9 @@ export async function navToSynteny({
   let rFeatStart: number
   let rFeatEnd: number
 
-  if (reg && cigar) {
-    const regStart = reg.start
-    const regEnd = reg.end
+  if (region && cigar) {
+    const regStart = region.start
+    const regEnd = region.end
     const p = parseCigar(cigar)
     const [fStartX, mStartX] = findPosInCigar(p, regStart - featStart)
     const [fEndX, mEndX] = findPosInCigar(p, regEnd - featStart)
