+++ plugins/linear-comparative-view/src/SyntenyFeatureDetail/types.ts
@@ -1,0 +1,10 @@
+import type { SimpleFeatureSerialized } from '@jbrowse/core/util'
+
+export interface SyntenyFeatureDetailModel {
+  trackId: string
+  featureData: SimpleFeatureSerialized
+  level?: number
+  view: {
+    type: string
+  }
+}
